+++ zio-http/src/test/scala/zio/http/ClientHttpsSpec.scala
@@ -1,13 +1,12 @@
-package zio.http.service
+package zio.http
 
+import zio.durationInt
 import zio.test.Assertion.{anything, equalTo, fails, isSubtype}
 import zio.test.TestAspect.{ignore, timeout}
 import zio.test.{ZIOSpecDefault, assertZIO}
-import zio.{Scope, durationInt}
 
 import zio.http.model.Status
-import zio.http.netty.client.{NettyClientDriver, NettyConnectionPool}
-import zio.http.{Client, ClientConfig, ClientSSLConfig}
+import zio.http.netty.client.NettyClientDriver
 
 import io.netty.handler.codec.DecoderException
