+++ zio-http/src/test/scala/zio/http/ClientProxySpec.scala
@@ -1,4 +1,4 @@
-package zio.http.service
+package zio.http
 
 import java.net.ConnectException
 
@@ -7,11 +7,10 @@ import zio.test.TestAspect.{sequential, timeout}
 import zio.test._
 import zio.{Scope, ZIO, durationInt}
 
-import zio.http._
 import zio.http.internal.{DynamicServer, HttpRunnableSpec, severTestLayer}
 import zio.http.middleware.Auth.Credentials
 import zio.http.model._
-import zio.http.netty.client.{NettyClientDriver, NettyConnectionPool}
+import zio.http.netty.client.NettyClientDriver
 
 object ClientProxySpec extends HttpRunnableSpec {
