+++ zio-http/src/test/scala/zio/http/ClientSpec.scala
@@ -1,15 +1,14 @@
-package zio.http.service
+package zio.http
 
 import java.net.ConnectException
 
-import zio.test.Assertion.{isLeft, _}
+import zio.test.Assertion._
 import zio.test.TestAspect.{sequential, timeout}
 import zio.test.assertZIO
 import zio.{Scope, durationInt}
 
 import zio.stream.ZStream
 
-import zio.http._
 import zio.http.internal.{DynamicServer, HttpRunnableSpec, severTestLayer}
 import zio.http.model._
