+++ zio-http/src/test/scala/zio/http/StaticServerSpec.scala
@@ -1,4 +1,4 @@
-package zio.http.service
+package zio.http
 
 import zio.test.Assertion.equalTo
 import zio.test.TestAspect.timeout
@@ -6,7 +6,6 @@ import zio.test.{Gen, TestEnvironment, assertTrue, assertZIO, checkAll}
 import zio.{Exit, Scope, ZIO, durationInt}
 
 import zio.http.Middleware.cors
-import zio.http._
 import zio.http.internal.{DynamicServer, HttpGen, HttpRunnableSpec, severTestLayer}
 import zio.http.middleware.Cors.CorsConfig
 import zio.http.model._
