+++ zio-http/src/test/scala/zio/http/WebSocketSpec.scala
@@ -1,4 +1,4 @@
-package zio.http.service
+package zio.http
 
 import zio._
 import zio.test.Assertion.equalTo
@@ -7,7 +7,6 @@ import zio.test.{TestClock, assertCompletes, assertTrue, assertZIO, testClock}
 
 import zio.http.ChannelEvent.UserEvent.HandshakeComplete
 import zio.http.ChannelEvent.{ChannelRead, ChannelUnregistered, UserEventTriggered}
-import zio.http._
 import zio.http.internal.{DynamicServer, HttpRunnableSpec, severTestLayer}
 import zio.http.model.{Headers, Status}
 import zio.http.socket.{WebSocketChannelEvent, WebSocketFrame}
